requirements.txt
numpy==1.19.5
opencv_python==4.5.3.56
matplotlib==3.4.2
Pillow==8.3.2
PySimpleGUI==4.49.0
